python_api/requirements.txt
fastapi
uvicorn
python-multipart
yt-dlp
PyYAML
huggingface_hub
numpy
soundfile
openai-whisper
torch
torchaudio
